server: Factor out ThoughtspaceExtension hocuspocus server extension.
